+++ src/main/java/com/buuz135/industrial/utils/compat/baubles/MeatFeederBauble.java
@@ -1,0 +1,91 @@
+package com.buuz135.industrial.utils.compat.baubles;
+
+import baubles.api.BaubleType;
+import baubles.api.IBauble;
+import baubles.api.cap.BaublesCapabilities;
+import com.buuz135.industrial.proxy.ItemRegistry;
+import com.buuz135.industrial.proxy.event.MeatFeederTickHandler;
+import com.buuz135.industrial.utils.Reference;
+import net.minecraft.entity.EntityLivingBase;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.NBTTagCompound;
+import net.minecraft.util.EnumFacing;
+import net.minecraft.util.ResourceLocation;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.common.capabilities.ICapabilitySerializable;
+import net.minecraftforge.event.AttachCapabilitiesEvent;
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
+public class MeatFeederBauble implements IBauble, ICapabilitySerializable<NBTTagCompound> {
+
+    @Override
+    public BaubleType getBaubleType(ItemStack itemStack) {
+        return BaubleType.HEAD;
+    }
+
+    @Override
+    public void onWornTick(ItemStack itemstack, EntityLivingBase player) {
+        if (player instanceof EntityPlayer) MeatFeederTickHandler.meatTick(itemstack, (EntityPlayer) player);
+    }
+
+    @Override
+    public void onEquipped(ItemStack itemstack, EntityLivingBase player) {
+
+    }
+
+    @Override
+    public void onUnequipped(ItemStack itemstack, EntityLivingBase player) {
+
+    }
+
+    @Override
+    public boolean canEquip(ItemStack itemstack, EntityLivingBase player) {
+        return true;
+    }
+
+    @Override
+    public boolean canUnequip(ItemStack itemstack, EntityLivingBase player) {
+        return true;
+    }
+
+    @Override
+    public boolean willAutoSync(ItemStack itemstack, EntityLivingBase player) {
+        return false;
+    }
+
+    @Override
+    public boolean hasCapability(@Nonnull Capability<?> capability, @Nullable EnumFacing facing) {
+        return capability == BaublesCapabilities.CAPABILITY_ITEM_BAUBLE;
+    }
+
+    @Nullable
+    @Override
+    public <T> T getCapability(@Nonnull Capability<T> capability, @Nullable EnumFacing facing) {
+        return capability == BaublesCapabilities.CAPABILITY_ITEM_BAUBLE ? (T) this : null;
+    }
+
+    @Override
+    public NBTTagCompound serializeNBT() {
+        return new NBTTagCompound();
+    }
+
+    @Override
+    public void deserializeNBT(NBTTagCompound nbt) {
+
+    }
+
+    public static class Event {
+
+        @SubscribeEvent
+        public void onAttachCapability(AttachCapabilitiesEvent<ItemStack> event) {
+            if (event.getObject().getItem().equals(ItemRegistry.meatFeederItem)) {
+                event.addCapability(new ResourceLocation(Reference.MOD_ID, "meat_feeder"), new MeatFeederBauble());
+            }
+        }
+
+    }
+}